Move derive macro impl to submodule
